consumer/requirements.txt
pika==1.3.2
PyMySQL==1.1.0
requests==2.31.0
redis==5.0.1
backoff==2.2.1
cryptography==42.0.4
python-dotenv==1.0.1
